run for IV, lots of helper methods, new controller interface

+++ src/cs3500/animator/model/actions/IAction.java
@@ -56,5 +56,5 @@ public interface IAction {
    * @param   speed The speed of the animation.
    * @return  The SVG description.
    */
-  String getSVGDescription(int speed);
+  String getSVGDescription(double speed);
 }
